refactor: FollowingTweetsRepositoryのRequestクラスとResponseクラスを分離
