+++ lib/models/repositories/following_tweets/following_tweets_api_request.dart
@@ -1,0 +1,11 @@
+class FollowingTweetsApiRequest {
+  final String eventId;
+
+  FollowingTweetsApiRequest({
+    required this.eventId,
+  });
+
+  Map<String, String> toParams() => {
+    "event_id": this.eventId,
+  };
+}
