+++ lib/models/repositories/following_tweets/following_tweets_api_response.dart
@@ -1,0 +1,14 @@
+import 'package:event_follow/models/entities/following_tweet.dart';
+
+class FollowingTweetsApiResponse {
+  FollowingTweetsApiResponse({
+    required this.tweets,
+  });
+
+  List<FollowingTweet> tweets;
+
+  factory FollowingTweetsApiResponse.fromJson(List<dynamic> json) =>
+      FollowingTweetsApiResponse(
+        tweets: List<FollowingTweet>.from(json.map((x) => FollowingTweet.fromJson(x))),
+      );
+}
