+++ lib/models/repositories/following_tweets/following_tweets_repository.dart
@@ -1,8 +1,9 @@
 import 'dart:convert';
 import 'dart:io';
-import 'package:event_follow/models/entities/entities.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:http/http.dart' as http;
+import 'following_tweets_api_request.dart';
+import 'following_tweets_api_response.dart';
 
 final followingTweetsRepositoryProvider = Provider.autoDispose.family<FollowingTweetsRepository, dynamic>(
         (ref, getOrGenerateIdToken) => FollowingTweetsRepository(getOrGenerateIdToken: getOrGenerateIdToken));
@@ -12,7 +13,7 @@ class FollowingTweetsRepository {
 
   FollowingTweetsRepository({required this.getOrGenerateIdToken});
 
-  Future<FollowingTweetsApiResults> requestFollowingTweetsApi(
+  Future<FollowingTweetsApiResponse> requestFollowingTweetsApi(
       {required FollowingTweetsApiRequest request}) async {
     final url = Uri.https("event-follow-front.herokuapp.com",
         "/api/following_tweets", request.toParams());
@@ -25,31 +26,6 @@ class FollowingTweetsRepository {
       },
     );
 
-    return FollowingTweetsApiResults.fromJson(json.decode(response.body));
+    return FollowingTweetsApiResponse.fromJson(json.decode(response.body));
   }
-}
-
-class FollowingTweetsApiRequest {
-  final String eventId;
-
-  FollowingTweetsApiRequest({
-    required this.eventId,
-  });
-
-  Map<String, String> toParams() => {
-    "event_id": this.eventId,
-  };
-}
-
-class FollowingTweetsApiResults {
-  FollowingTweetsApiResults({
-    required this.tweets,
-  });
-
-  List<FollowingTweet> tweets;
-
-  factory FollowingTweetsApiResults.fromJson(List<dynamic> json) =>
-      FollowingTweetsApiResults(
-        tweets: List<FollowingTweet>.from(json.map((x) => FollowingTweet.fromJson(x))),
-      );
 }
