+++ lib/pages/events_pages/friends_footer.dart
@@ -2,7 +2,7 @@ import 'package:event_follow/models/controllers/following_tweets/following_tweet
 import 'package:event_follow/models/controllers/friendships_controller/friendships_controller.dart';
 import 'package:event_follow/models/entities/event.dart';
 import 'package:event_follow/models/entities/event_extra.dart';
-import 'package:event_follow/models/models.dart';
+import 'package:event_follow/models/repositories/following_tweets/following_tweets_api_request.dart';
 import 'package:event_follow/models/repositories/friendships/friendships_api_request.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
